requirements.txt
aiosqlite==0.17.0
anyio==3.6.1
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
click==8.1.3
coverage==6.4.4
fastapi==0.82.0
graphql-core==3.2.1
greenlet==1.1.3
h11==0.13.0
idna==3.3
iniconfig==1.1.1
kink==0.6.5
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pytest==7.1.3
pytest-asyncio==0.19.0
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.19.1
strawberry-graphql==0.128.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
websockets==10.3
